[#132][hub] refactor: 패키지 이동
