+++ hub/src/main/java/com/faster/hub/app/hub/application/usecase/HubService.java
@@ -1,16 +1,16 @@
 package com.faster.hub.app.hub.application.usecase;
 
-import com.faster.hub.app.hub.application.dto.request.GetPathApplicationRequestDto;
-import com.faster.hub.app.hub.application.dto.request.SaveHubApplicationRequestDto;
-import com.faster.hub.app.hub.application.dto.response.GetHubsApplicationResponseDto;
-import com.faster.hub.app.hub.application.dto.response.GetPathsApplicationResponseDto;
-import com.faster.hub.app.hub.application.dto.response.SaveHubApplicationResponseDto;
-import com.faster.hub.app.hub.application.dto.response.GetHubApplicationResponseDto;
-import com.faster.hub.app.hub.application.dto.request.DeleteHubApplicationRequestDto;
-import com.faster.hub.app.hub.application.dto.request.UpdateHubApplicationRequestDto;
+import com.faster.hub.app.hub.application.usecase.dto.request.GetPathApplicationRequestDto;
+import com.faster.hub.app.hub.application.usecase.dto.request.SaveHubApplicationRequestDto;
+import com.faster.hub.app.hub.application.usecase.dto.response.GetHubsApplicationResponseDto;
+import com.faster.hub.app.hub.application.usecase.dto.response.GetPathsApplicationResponseDto;
+import com.faster.hub.app.hub.application.usecase.dto.response.SaveHubApplicationResponseDto;
+import com.faster.hub.app.hub.application.usecase.dto.response.GetHubApplicationResponseDto;
+import com.faster.hub.app.hub.application.usecase.dto.request.DeleteHubApplicationRequestDto;
+import com.faster.hub.app.hub.application.usecase.dto.request.UpdateHubApplicationRequestDto;
 import java.util.List;
 import java.util.UUID;
-import com.faster.hub.app.hub.application.dto.response.UpdateHubApplicationResponseDto;
+import com.faster.hub.app.hub.application.usecase.dto.response.UpdateHubApplicationResponseDto;
 
 public interface HubService {
