+++ hub/src/main/java/com/faster/hub/app/hub/application/usecase/HubServiceImpl.java
@@ -2,15 +2,15 @@ package com.faster.hub.app.hub.application.usecase;
 
 import com.common.exception.CustomException;
 import com.faster.hub.app.global.exception.HubErrorCode;
-import com.faster.hub.app.hub.application.dto.request.DeleteHubApplicationRequestDto;
-import com.faster.hub.app.hub.application.dto.request.GetPathApplicationRequestDto;
-import com.faster.hub.app.hub.application.dto.request.SaveHubApplicationRequestDto;
-import com.faster.hub.app.hub.application.dto.request.UpdateHubApplicationRequestDto;
-import com.faster.hub.app.hub.application.dto.response.GetHubApplicationResponseDto;
-import com.faster.hub.app.hub.application.dto.response.GetHubsApplicationResponseDto;
-import com.faster.hub.app.hub.application.dto.response.GetPathsApplicationResponseDto;
-import com.faster.hub.app.hub.application.dto.response.SaveHubApplicationResponseDto;
-import com.faster.hub.app.hub.application.dto.response.UpdateHubApplicationResponseDto;
+import com.faster.hub.app.hub.application.usecase.dto.request.DeleteHubApplicationRequestDto;
+import com.faster.hub.app.hub.application.usecase.dto.request.GetPathApplicationRequestDto;
+import com.faster.hub.app.hub.application.usecase.dto.request.SaveHubApplicationRequestDto;
+import com.faster.hub.app.hub.application.usecase.dto.request.UpdateHubApplicationRequestDto;
+import com.faster.hub.app.hub.application.usecase.dto.response.GetHubApplicationResponseDto;
+import com.faster.hub.app.hub.application.usecase.dto.response.GetHubsApplicationResponseDto;
+import com.faster.hub.app.hub.application.usecase.dto.response.GetPathsApplicationResponseDto;
+import com.faster.hub.app.hub.application.usecase.dto.response.SaveHubApplicationResponseDto;
+import com.faster.hub.app.hub.application.usecase.dto.response.UpdateHubApplicationResponseDto;
 import com.faster.hub.app.hub.domain.repository.HubRepository;
 import java.util.List;
 import java.util.UUID;
