+++ hub/src/main/java/com/faster/hub/app/hub/application/usecase/PathFinder.java
@@ -1,6 +1,6 @@
 package com.faster.hub.app.hub.application.usecase;
 
-import com.faster.hub.app.hub.application.dto.response.GetPathsApplicationResponseDto;
+import com.faster.hub.app.hub.application.usecase.dto.response.GetPathsApplicationResponseDto;
 import com.faster.hub.app.hub.domain.entity.Hub;
 import java.util.List;
 import java.util.UUID;
