+++ hub/src/main/java/com/faster/hub/app/hub/presentation/dto/response/GetPathsResponseDto.java
@@ -1,6 +1,6 @@
 package com.faster.hub.app.hub.presentation.dto.response;
 
-import com.faster.hub.app.hub.application.dto.response.GetPathsApplicationResponseDto;
+import com.faster.hub.app.hub.application.usecase.dto.response.GetPathsApplicationResponseDto;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
